% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{a}
\alias{a}
\title{Function returning the adjusting parameter a slot of an S4 object}
\usage{
a(obj)
}
\arguments{
\item{obj}{object of class randPar}
}
\description{
Function returning the adjusting parameter a slot of an S4 object
}
