% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcdPar.R
\name{abcdRand}
\alias{abcdRand}
\title{Sampling algorithm for abcd}
\usage{
abcdRand(N, a, K = 2)
}
\arguments{
\item{N}{integer for the total sample size of the trial.}

\item{a}{nonnegative parameter which my be adjusted according to how strongly it is 
desired to balance the experiment. a = 0 gives the complete randomization, while the
assignments become more deterministic as \code{a} increases.}

\item{K}{number of treatment groups (e.g. K=2 if we compare one experimental
against one control treatment).}
}
\value{
A vector with the allocation sequence for a clinical trial. 
It will contain a zero (resp. 1) at position \code{i}, when patient \code{i}
is allocated to treatment A (resp. B).
}
\description{
Sampling algorithm for abcd
}
\references{
A. B. Antognini and A. Giovagnoli (2004) A new 'biased coin design' for the sequential 
allocation of two treatments.
\emph{Journal of the Royal Statistical Society. Series C (Applied Statistics)}
\strong{53}, No. 4, 651-664
}

