% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbcdPar.R
\name{bbcdRand}
\alias{bbcdRand}
\title{Sampling algorithm for bbcd}
\usage{
bbcdRand(N, a, K = 2)
}
\arguments{
\item{N}{integer for the total sample size of the trial.}

\item{a}{nonnegative parameter which controls the degree of randomness:
For decreasing \code{a} the allocations become deterministic, while for increasing
\code{a} the randomization procedure tends to complete randomization.}

\item{K}{number of treatment groups (e.g. K=2 if we compare one experimental
against one control treatment).}
}
\value{
A vector with the allocation sequence for a clinical trial.
It will contain a zero (resp. 1) at position \code{i}, when patient \code{i}
is allocated to treatment A (resp. B).
}
\description{
Sampling algorithm for bbcd
}
\references{
A. B. Antognini and Maroussa Zagoraiou (2014) Balance and randomness in sequential
clinical trials: the dominant biased coin design.
\emph{Pharmaceutical Statistics}
\strong{13(2)}, 119-127
}
\keyword{internal}
