% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randPar.R
\name{blocks}
\alias{blocks}
\title{Function returning the block slot of an S4 object}
\usage{
blocks(obj)
}
\arguments{
\item{obj}{object of class pbrPAr}
}
\description{
Function returning the block slot of an S4 object
}

