% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsdPar.R
\name{bsdRand}
\alias{bsdRand}
\title{Sampling algorithm for BSD}
\usage{
bsdRand(N, mti, K = 2)
}
\arguments{
\item{N}{integer for the total sample size of the trial.}

\item{mti}{maximum tolerated imbalance in patient numbers during the trial.}

\item{K}{number of treatment groups (e.g. K=2 if we compare one experimental
against one control treatment).}
}
\value{
A vector with the allocation sequence for a clinical trial. 
It will contain a zero (resp. 1) at position \code{i}, when patient \code{i}
is allocated to treatment A (resp. B).
}
\description{
Sampling algorithm for BSD
}
\references{
J. F. Soares and C. F. Jeff Wu (1983) Some Restricted Randomization Rules in
Sequential Designs. Comm. in Stat., 12, 2017-34.
}
