% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{$,assessment-method}
\alias{$,assessment-method}
\title{Method defining the $ operator for the assessment class}
\usage{
\S4method{$}{assessment}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the assessment class
}
\keyword{internal}
