% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\docType{methods}
\name{$,comparison-method}
\alias{$,comparison-method}
\title{Method defining the $ operator for the assessemnt class}
\usage{
\S4method{$}{comparison}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the assessemnt class
}
