% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desFunc.R
\docType{methods}
\name{$,desFunc-method}
\alias{$,desFunc-method}
\title{Method defining the $ operator for the desFunc class}
\usage{
\S4method{$}{desFunc}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the desFunc class
}
