% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\docType{methods}
\name{$,endpoint-method}
\alias{$,endpoint-method}
\title{Method defining the $ operator for the endpoint class}
\usage{
\S4method{$}{endpoint}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the endpoint class
}
