% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\docType{methods}
\name{$,evaluation-method}
\alias{$,evaluation-method}
\title{Method defining the $ operator for the evaluation class}
\usage{
\S4method{$}{evaluation}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the evaluation class
}
