% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/issue.R
\docType{methods}
\name{$,issue-method}
\alias{$,issue-method}
\title{Method defining the $ operator for the issue class}
\usage{
\S4method{$}{issue}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the issue class
}

