% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probUnDes.R
\docType{methods}
\name{$,probUnDesirable-method}
\alias{$,probUnDesirable-method}
\title{Method defining the $ operator for the probUnDesirable class}
\usage{
\S4method{$}{probUnDesirable}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the probUnDesirable class
}
