% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randPar.R
\docType{methods}
\name{$,randPar-method}
\alias{$,randPar-method}
\title{Method defining the $ operator for the randPar class}
\usage{
\S4method{$}{randPar}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the randPar class
}

