% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSeq.R
\docType{methods}
\name{$,randSeq-method}
\alias{$,randSeq-method}
\title{Method defining the $ operator for the randSeq class}
\usage{
\S4method{$}{randSeq}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the randSeq class
}

