% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chenPar.R
\name{chenRand}
\alias{chenRand}
\title{Representing Chen's Design}
\usage{
chenRand(N, mti, p, K = 2)
}
\arguments{
\item{N}{integer for the total sample size of the trial.}

\item{mti}{maximum tolerated imbalance in patient numbers during the trial.}

\item{p}{success probability of the biased coin (e.g. in Efron's Biased Coin
Design).}

\item{K}{number of treatment groups (e.g. K=2 if we compare one experimental
against one control treatment).}
}
\value{
A vector with the allocation sequence for a clinical trial. 
It will contain a zero (resp. 1) at position \code{i}, when patient \code{i}
is allocated to treatment A (resp. B).
}
\description{
Representing Chen's Design
}
\references{
Chen Yung-Pin (1999) Biased coin design with imbalance tolerance.
Comm. in Stat., 15, 953-975.
}
\keyword{internal}
