% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randPar.R
\name{coin}
\alias{coin}
\title{Function returning the coin slot of an S4 object}
\usage{
coin(obj)
}
\arguments{
\item{obj}{object extending class randPar or randSeq}
}
\description{
Function returning the coin slot of an S4 object
}

