% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedBias.R
\name{combineBias}
\alias{combineBias}
\title{Combined additive bias criterion}
\usage{
combineBias(selBias, chronBias)
}
\arguments{
\item{selBias}{object of class \code{selBias}}

\item{chronBias}{object of class \code{chronBias}}
}
\description{
This class combines a \code{selBias} object and a \code{chronBias} object
to a new object. In the analysis within the new object
the two types of bias are treated as additive.
effect.
}
\examples{
chronBias <- chronBias(type="linT", theta=1, method="sim")
selBias <- selBias(type="CS", eta=1, method="sim")
combineBias(selBias, chronBias)

}
\seealso{
Other issues: \code{\link{chronBias}},
  \code{\link{corGuess}}, \code{\link{imbal}},
  \code{\link{issue}}, \code{\link{selBias}},
  \code{\link{setPower}}
}
