% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{createSeq}
\alias{createSeq}
\title{Query to create a randomization sequence of a particular randomization procedure}
\usage{
createSeq(file)
}
\arguments{
\item{file}{A connection, or a character string naming the file to write to.}
}
\value{
an object \code{Param}, which is available
}
\description{
This function is a query to create an corresponding randomization sequence
for a two-armed clinical trial. If
\code{file} is defined, the generated sequence is automatically saved to the
corresponding path.
}

