% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crPar.R
\name{crPar}
\alias{crPar}
\title{Representing Complete Randomization}
\usage{
crPar(N, K = 2, ratio = rep(1, K), groups = LETTERS[1:K])
}
\arguments{
\item{N}{integer for the total sample size of the trial.}

\item{K}{number of treatment groups (e.g. K=2 if we compare one experimental
against one control treatment).}

\item{ratio}{vector of length \code{K}. The total sample number \code{N} and
all used block lengths (\code{bc}) have to be divisible by \code{sum(ratio)}.}

\item{groups}{character vector of labels for the different treatments.}
}
\value{
\code{S4} object of the class \code{crPar}.
}
\description{
Represents the randomization procedure Complete Randomization.
}
\details{
Toss a fair coin \code{N} times in case \code{K=2} and assign the treatments according to the result of the coin. In case of \code{K>2}, replace the coin by a die with \code{K} sides.
}
\references{
W. F. Rosenberger and J. M. Lachin (2002) \emph{Randomization in Clinical Trials}.
Wiley.
}
\seealso{
Other randomization procedures: \code{\link{abcdPar}},
  \code{\link{bbcdPar}}, \code{\link{bsdPar}},
  \code{\link{chenPar}}, \code{\link{createParam}},
  \code{\link{ebcPar}}, \code{\link{gbcdPar}},
  \code{\link{hadaPar}}, \code{\link{mpPar}},
  \code{\link{pbrPar}}, \code{\link{rarPar}},
  \code{\link{rpbrPar}}, \code{\link{rtbdPar}},
  \code{\link{tbdPar}}, \code{\link{udPar}}
}
