% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{doublyF_opt}
\alias{doublyF_opt}
\title{Distribution function of the non central F-distribution}
\usage{
doublyF_opt(x, df1, df2, lambda1, lambda2, acc = 100, ex = 5)
}
\arguments{
\item{x}{quantile, a single numeric value}

\item{df1}{first degree of freedom, a single integer value}

\item{df2}{second degree of freedom, a single integer value}

\item{lambda1}{first non centrality parameter, a single numeric value}

\item{lambda2}{second non centrality parameter, a single numeric value}

\item{acc}{accuracy; last index of the approximation of the infinite sum}

\item{ex}{exactness; break early, if the summands are smaller than \code{10^(-ex)}}
}
\value{
Probability of observing a value larger than x
}
\description{
Distribution function of the non central F-distribution
}
