% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{doublyF_value}
\alias{doublyF_value}
\title{Rejection probability for one sequence in the presence of selection bias}
\usage{
doublyF_value(R, bias, endp)
}
\arguments{
\item{R}{object of type randSeq representing the randomization procedure}

\item{bias}{selection or chronological bias object, containing eta(or theta) and alpha}

\item{endp}{endpoint object, containing mu and sigma}
}
\value{
data frame containing the non centrality parameters, the degrees of
freedom, the \code{1-alpha} quantile corresponding to the central f distribution and
a numeric value for probability of false rejection of the null
hypothesis of no difference in the presence of selection bias.
}
\description{
Rejection probability for one sequence in the presence of selection bias
}
\keyword{internal}
