% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{doublyF_values}
\alias{doublyF_values}
\title{Rejection probability in case of selection bias in multi-arm trials}
\usage{
doublyF_values(randSeq, bias, endp)
}
\arguments{
\item{randSeq}{the object containing the randomization sequences}

\item{bias}{selection bias object, containing eta and alpha}

\item{endp}{endpoint object, containing mu and sigma}
}
\value{
data frame with the sequences and their corresponding ncps and rejection
probabilities.
}
\description{
calculates the non-centrality parameters of the F-distribution under third
order selection bias.
}
\keyword{internal}
