% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyT.R
\name{doublyT}
\alias{doublyT}
\title{Approximation of the distribution function of the doubly noncentral t-distribution}
\usage{
doublyT(x, df, lambda1, lambda2, lb = 0, ub)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{df}{degrees of freedom (i.a. \code{N-2}).}

\item{lb}{lower bound for the starting value of the poisson distribution.}

\item{ub}{upper bound for the last value of the poisson distribution.}
}
\value{
Distribution value of the doubly noncentral t-distribution at \code{x}.
}
\description{
Computes the value of the distribution function of the doubly noncentral t-distribution at \code{x}.
}
\keyword{internal}
