% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyT.R
\name{genNcps_new}
\alias{genNcps_new}
\title{Calculation of the NCPs of each randomization sequence for the doubly noncentral t-distribution}
\usage{
genNcps_new(randSeq, bias, endp, weight = FALSE)
}
\arguments{
\item{randSeq}{a list of randSeq(rCrSeq or others) with possible varying N's}

\item{bias}{a list of biases - corresponding to the different randSeq's}

\item{endp}{object of the class endpoint.}

\item{weight}{if set to TRUE the weight will be set to 1, according to the paper}
}
\value{
a list containing the noncentrality parameters delta and lambda of all randomization sequences.
}
\description{
Computes the noncentrality parameters delta and lambda for the doubly noncentral t-distribution of each randomization sequence.
}
