% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyT.R
\name{get_p_values_new}
\alias{get_p_values_new}
\title{Calculating p values}
\usage{
get_p_values_new(delta, lambda, N, alpha = 0.05, df = sum(N - 2))
}
\arguments{
\item{delta}{The first noncentrality parameter}

\item{lambda}{The second noncentrality parameter}

\item{N}{the amount of patients in the trial}

\item{alpha}{significance level}

\item{df}{degrees of freedom}
}
\value{
a p value
}
\description{
Computes the p values based on the noncentrality parameters delta and lambda for the doubly noncentral t-distribution
}
