% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{getbiasCS1}
\alias{getbiasCS1}
\title{Biasing Policy for a Group of Favoured Treatments}
\usage{
getbiasCS1(R, K, pref)
}
\arguments{
\item{R}{randomization sequence, a integer vector with entries 1, \dots, K, of length N}

\item{K}{number of treatment groups, a single integer value}

\item{pref}{preferred groups for the guessing}
}
\value{
vector with the selection bias for each patient
}
\description{
Calculate vector with the selection bias for each patient
}
