% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{getbiasCS2}
\alias{getbiasCS2}
\title{Biasing Policy for Avoiding the Placebo Treatment}
\usage{
getbiasCS2(R, K, avoid)
}
\arguments{
\item{R}{randomization sequence, a integer vector with entries 1, \dots, K, of length N}

\item{K}{number of treatment groups, a single integer value}

\item{avoid}{avoided groups for the guessing}
}
\value{
vector with the selection bias for each patient
}
\description{
Calculate vector with the selection bias for each patient
}
\keyword{internal}
