% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corGuess.R
\name{getCorGuesses}
\alias{getCorGuesses}
\title{Matrix of the guesses of the investigator}
\usage{
getCorGuesses(randSeq, guessing)
}
\arguments{
\item{randSeq}{object of the class randSeq.}

\item{guessing}{object of the class corGuess.}
}
\value{
Matrix of the guesses of the investigator following the specified guessing
 strategy. No guess is abbreviated with \code{"nG"}.
}
\description{
Calculates the guesses of the investigator of a randomization list following
the specified guessing strategy.
}
\examples{
myPar <- bsdPar(10, 2)
M <- genSeq(myPar, 2)
type <- corGuess("CS")
getCorGuesses(M, type)

}
