% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bsdPar.R, R/bsdSeq.R, R/crPar.R, R/crSeq.R, R/ebcPar.R, R/ebcSeq.R, R/getDesign.R, R/hadaPar.R, R/hadaSeq.R, R/mpPar.R, R/mpSeq.R, R/pbrPar.R, R/pbrSeq.R, R/rarPar.R, R/rarSeq.R, R/rpbrPar.R, R/rpbrSeq.R, R/rtbdPar.R, R/rtbdSeq.R, R/tbdPar.R, R/tbdSeq.R, R/udPar.R, R/udSeq.R
\docType{methods}
\name{getDesign,bsdPar-method}
\alias{getDesign}
\alias{getDesign,bsdPar-method}
\alias{getDesign,bsdSeq-method}
\alias{getDesign,crPar-method}
\alias{getDesign,crSeq-method}
\alias{getDesign,ebcPar-method}
\alias{getDesign,ebcSeq-method}
\alias{getDesign,hadaPar-method}
\alias{getDesign,hadaSeq-method}
\alias{getDesign,mpPar-method}
\alias{getDesign,mpSeq-method}
\alias{getDesign,pbrPar-method}
\alias{getDesign,pbrSeq-method}
\alias{getDesign,rRpbrSeq-method}
\alias{getDesign,rRtbdSeq-method}
\alias{getDesign,rarPar-method}
\alias{getDesign,rarSeq-method}
\alias{getDesign,rpbrPar-method}
\alias{getDesign,rtbdPar-method}
\alias{getDesign,tbdPar-method}
\alias{getDesign,tbdSeq-method}
\alias{getDesign,udPar-method}
\alias{getDesign,udSeq-method}
\title{Design of a randomization procedure}
\usage{
\S4method{getDesign}{bsdPar}(obj)

\S4method{getDesign}{bsdSeq}(obj)

\S4method{getDesign}{crPar}(obj)

\S4method{getDesign}{crSeq}(obj)

\S4method{getDesign}{ebcPar}(obj)

\S4method{getDesign}{ebcSeq}(obj)

getDesign(obj)

\S4method{getDesign}{hadaPar}(obj)

\S4method{getDesign}{hadaSeq}(obj)

\S4method{getDesign}{mpPar}(obj)

\S4method{getDesign}{mpSeq}(obj)

\S4method{getDesign}{pbrPar}(obj)

\S4method{getDesign}{pbrSeq}(obj)

\S4method{getDesign}{rarPar}(obj)

\S4method{getDesign}{rarSeq}(obj)

\S4method{getDesign}{rpbrPar}(obj)

\S4method{getDesign}{rRpbrSeq}(obj)

\S4method{getDesign}{rtbdPar}(obj)

\S4method{getDesign}{rRtbdSeq}(obj)

\S4method{getDesign}{tbdPar}(obj)

\S4method{getDesign}{tbdSeq}(obj)

\S4method{getDesign}{udPar}(obj)

\S4method{getDesign}{udSeq}(obj)
}
\arguments{
\item{obj}{object of the class \code{randSeq} or \code{randPar}.}
}
\description{
Generates a \code{character} vector which specifies the used randomization method
}

