% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSeq.R
\name{getRandomizationList}
\alias{getRandomizationList}
\alias{getRandList}
\title{Accessor function for the randomization list}
\usage{
getRandList(obj)
}
\arguments{
\item{obj}{object specifying the randomization procedure, see \code{\link{randPar}}
or \code{\link{createParam}}.}
}
\value{
A matrix with all randomization sequences of a \code{S4} object
}
\description{
Get the randomization list coded in its groups.
}
\examples{
myPar <- bsdPar(10, 2)
M <- genSeq(myPar, 2)
getRandList(M)
}
