% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{hatMatrix}
\alias{hatMatrix}
\title{Calculate hat matrix}
\usage{
hatMatrix(M)
}
\arguments{
\item{M}{Design Matrix representing the randomization sequence}
}
\value{
Hat matrix, i.e. \code{M * (M^T M)^{-1} M^T}
}
\description{
Calculate hat matrix
}
