% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{lambda}
\alias{lambda}
\title{Method returning the rate parameter of an expEndp S4 object}
\usage{
lambda(obj)
}
\arguments{
\item{obj}{object of class expEndp}
}
\description{
Method returning the rate parameter of an expEndp S4 object
}
