% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{lambda1}
\alias{lambda1}
\title{Calculate first non centrality parameter of the doubly non central F-distribution}
\usage{
lambda1(H, EY)
}
\arguments{
\item{H}{Hat Matrix}

\item{EY}{(Biased) expectation of the responses}
}
\value{
First non centrality parameter, a single numeric value
}
\description{
Calculate first non centrality parameter of the doubly non central F-distribution
}
