% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{lambda2}
\alias{lambda2}
\title{Calculate second non centrality parameter of the doubly non central F-distribution}
\usage{
lambda2(H, EY)
}
\arguments{
\item{H}{Hat Matrix}

\item{EY}{(Biased) expectation of the responses}
}
\value{
Second non centrality parameter, a single numeric value
}
\description{
Calculate second non centrality parameter of the doubly non central F-distribution
}
