% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRankDistribution.R
\name{logRankDecSim}
\alias{logRankDecSim}
\title{logRankDecSim}
\usage{
logRankDecSim(randSeq, bias, endp)
}
\arguments{
\item{randSeq}{object of the class randSeq.}

\item{bias}{object of the class bias.}

\item{endp}{object of the class endpoint.}
}
\value{
a simulated test decision for each randomization sequence.
}
\description{
logRankDecSim
}
\keyword{internal}
