% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRankDistribution.R
\name{logRankRejectionProb}
\alias{logRankRejectionProb}
\title{Approximation of the biased type I error probability (resp. power) of the log rank test}
\usage{
logRankRejectionProb(randSeq, bias, endp)
}
\arguments{
\item{randSeq}{object of the class randSeq.}

\item{bias}{object of the class bias.}

\item{endp}{object of the class endpoint.}
}
\value{
the rejection probability (type I error probability resp. power) of all randomization sequences.
}
\description{
Computes the biased type I error probability (resp. power) of the log rank test due to shifts in the expectation vectors
in both treatment groups.
}
\keyword{internal}
