% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{makeBiasedExpectation}
\alias{makeBiasedExpectation}
\title{Calculate Expectation vector}
\usage{
makeBiasedExpectation(R, mu, bias)
}
\arguments{
\item{R}{randomization sequence, an object of type randSeq}

\item{mu}{vector of length K containing the expectation values of groups 1, \dots, K}

\item{bias}{selection or chronological bias object, containing eta(or theta) and alpha}
}
\value{
vector of length N with the biased expectation for each patient.
}
\description{
Calculate Expectation vector
}
