% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randPar.R
\name{method}
\alias{method}
\title{Function returning the allocation ratio slot of an S4 object}
\usage{
method(obj)
}
\arguments{
\item{obj}{object of class randPar}
}
\description{
Function returning the allocation ratio slot of an S4 object
}

