% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randPar.R
\name{mti}
\alias{mti}
\title{Function returning the MTI slot of an S4 object}
\usage{
mti(obj)
}
\arguments{
\item{obj}{object of class bsdPar or mpPar}
}
\description{
Function returning the MTI slot of an S4 object
}

