% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{mu}
\alias{mu}
\title{Access the expectation value slot of a normEndp S4 object}
\usage{
mu(obj)
}
\arguments{
\item{obj}{object of class normEndp}
}
\description{
Access the expectation value slot of a normEndp S4 object
}
