% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{N}
\alias{N}
\title{Function returning the sample size slot of an S4 object}
\usage{
N(obj)
}
\arguments{
\item{obj}{object inheriting from randPar}
}
\value{
the sample size slot of an \code{S4} object
}
\description{
Function returning the sample size slot of an S4 object
}
