% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSeq.R
\name{plotSeq}
\alias{plotSeq}
\title{Sequence plotting}
\usage{
plotSeq(sequences, plotAllSeq = FALSE, emph = NA, rs = NA)
}
\arguments{
\item{sequences}{object of type randSeq}

\item{plotAllSeq}{logical. If \code{plotAllSeq=TRUE}, the complete set of
randomization sequences will be plotted in light gray.}

\item{emph}{integer indicating which sequence should be highlighted in blue.}

\item{rs}{vector of a randomization sequence that should be highlighted.}
}
\value{
A plot of all randomization sequences of a \code{randSeq} object.
}
\description{
Plot all randomization sequences of a randSeq object
}
