% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSeq.R
\docType{class}
\name{randSeq-class}
\alias{randSeq-class}
\title{An S4 Class for the representation of  randomization sequences}
\description{
This set of classes provides functionality of storing randomization
sequences of different randomization procedures along with the parameters 
representing the design.
}
\section{Slots}{

\describe{
\item{\code{N}}{total number of patients included in the trial}

\item{\code{M}}{matrix containing randomization sequences of length \code{N} in its
rows.}

\item{\code{K}}{number of treatment groups}

\item{\code{groups}}{character string of length K defining the names of the treatment groups}
}}

