% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{ratio}
\alias{ratio}
\title{Function returning the allocation ratio slot of an S4 object}
\usage{
ratio(obj)
}
\arguments{
\item{obj}{object of class randPar}
}
\description{
Function returning the allocation ratio slot of an S4 object
}
