% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{rho}
\alias{rho}
\title{Function returning the adjusting parameter rho slot of an S4 object}
\usage{
rho(obj)
}
\arguments{
\item{obj}{object of class randPar}
}
\description{
Function returning the adjusting parameter rho slot of an S4 object
}
