% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveAssess.R
\name{saveAssess}
\alias{saveAssess}
\title{Saving an assess object}
\usage{
saveAssess(obj, file = "assessObject.csv")
}
\arguments{
\item{obj}{object specifying the randomization procedure, i.e. an object of a
class.}

\item{file}{A connection, or a character string naming the file to write to.}
}
\value{
Creates a \code{.csv} data in the home folder.
}
\description{
Saves the fully information of an \code{assess} object in a \code{.csv} data
sheet.
}
\seealso{
Other saving functions: \code{\link{saveRand}}
}

