% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveRand.R
\name{saveRand}
\alias{saveRand}
\title{Saving a randomization lists}
\usage{
saveRand(obj, file = "randList.csv")
}
\arguments{
\item{obj}{object specifying the randomization procedure, i.e. an object of a
class.}

\item{file}{A connection, or a character string naming the file to write to.}
}
\value{
Creates a \code{.csv} data in the home folder.
}
\description{
Saves the parameters of a \code{randSeq} object in a \code{.csv} data
sheet.
}
\seealso{
Other saving functions: \code{\link{saveAssess}}
}

