% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{scale}
\alias{scale}
\title{Method returning the scale parameter of an survEndp S4 object}
\usage{
scale(obj)
}
\arguments{
\item{obj}{object of class survEndp}
}
\description{
Method returning the scale parameter of an survEndp S4 object
}
