% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{shape}
\alias{shape}
\title{Method returning the shape parameter of an survEndp S4 object}
\usage{
shape(obj)
}
\arguments{
\item{obj}{object of class survEndp}
}
\description{
Method returning the shape parameter of an survEndp S4 object
}
