% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{sigma}
\alias{sigma}
\title{Function returning the standard deviation slot of a normEndp S4 object}
\usage{
sigma(obj)
}
\arguments{
\item{obj}{object of class normEndp}
}
\description{
Function returning the standard deviation slot of a normEndp S4 object
}
