% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derFunc.R
\name{TV}
\alias{TV}
\title{Function returning the sample size slot of an S4 object}
\usage{
TV(obj)
}
\arguments{
\item{obj}{object inheriting from derFunc}
}
\description{
Function returning the sample size slot of an S4 object
}
