% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{type}
\alias{type}
\title{Get type of an object}
\usage{
type(obj)
}
\arguments{
\item{obj}{a bias object (i.e. S4 object inheriting from \code{bias})}
}
\value{
Character string specifying the type of bias \code{obj} represents, e.g. \code{"linT"} in 
case of chronological bias.
}
\description{
Accesses the type slot of an S4 object
}
