% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{blocks}
\alias{blocks}
\title{Function returning the block slot of an S4 object}
\usage{
blocks(obj)
}
\arguments{
\item{obj}{object of class pbrPar}
}
\value{
a vector with the lenghts of each block of a \code{pbrPar} object
}
\description{
Function returning the block slot of an S4 object
}
