% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derFunc.R
\docType{methods}
\name{$,derFunc-method}
\alias{$,derFunc-method}
\title{Method defining the $ operator for the derFunc class}
\usage{
\S4method{$}{derFunc}(x, name)
}
\arguments{
\item{x}{a variable \code{x}.}

\item{name}{name of a variable.}
}
\description{
Method defining the $ operator for the derFunc class
}
