% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyT.R
\name{genNcps}
\alias{genNcps}
\title{Calculation of the NCPs of each randomization sequence for the doubly noncentral t-distribution}
\usage{
genNcps(randSeq, bias, endp)
}
\arguments{
\item{randSeq}{object of the class randSeq.}

\item{bias}{object of the class bias.}

\item{endp}{object of the class endpoint.}
}
\value{
matrix containing the noncentrality parameters delta and lambda of all randomization sequences.
}
\description{
Computes the noncentrality parameters delta and lambda for the doubly noncentral t-distribution of each randomization sequence.
}
\keyword{internal}
