% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derringerLs.R, R/derringerRs.R, R/derringerTs.R, R/getDesFunc.R
\docType{methods}
\name{getDesFunc,derringerLs-method}
\alias{getDesFunc}
\alias{getDesFunc,derringerLs-method}
\alias{getDesFunc,derringerRs-method}
\alias{getDesFunc,derringerTs-method}
\title{Type of Desirability function}
\usage{
\S4method{getDesFunc}{derringerLs}(obj)

\S4method{getDesFunc}{derringerRs}(obj)

\S4method{getDesFunc}{derringerTs}(obj)

getDesFunc(obj)
}
\arguments{
\item{obj}{object of the class \code{desFunc}.}
}
\description{
Generates a \code{character} vector which specifies the used desirability function and
its parameters
}

