% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesign.R, R/pbrPar.R, R/rarPar.R,
%   R/rarSeq.R, R/abcdPar.R, R/abcdSeq.R, R/bbcdPar.R, R/bbcdSeq.R, R/ebcPar.R,
%   R/bsdPar.R, R/bsdSeq.R, R/chenPar.R, R/chenSeq.R, R/crPar.R, R/crSeq.R,
%   R/ebcSeq.R, R/gbcdPar.R, R/gbcdSeq.R, R/hadaPar.R, R/hadaSeq.R, R/mpPar.R,
%   R/mpSeq.R, R/pbrSeq.R, R/rtbdSeq.R, R/rpbrSeq.R, R/rpbrPar.R, R/tbdPar.R,
%   R/rtbdPar.R, R/tbdSeq.R, R/udPar.R, R/udSeq.R
\name{getDesign}
\alias{getDesign}
\alias{getDesign,pbrPar-method}
\alias{getDesign,rarPar-method}
\alias{getDesign,rarSeq-method}
\alias{getDesign,abcdPar-method}
\alias{getDesign,abcdSeq-method}
\alias{getDesign,bbcdPar-method}
\alias{getDesign,bbcdSeq-method}
\alias{getDesign,ebcPar-method}
\alias{getDesign,bsdPar-method}
\alias{getDesign,bsdSeq-method}
\alias{getDesign,chenPar-method}
\alias{getDesign,chenSeq-method}
\alias{getDesign,crPar-method}
\alias{getDesign,crSeq-method}
\alias{getDesign,ebcSeq-method}
\alias{getDesign,gbcdPar-method}
\alias{getDesign,gbcdSeq-method}
\alias{getDesign,hadaPar-method}
\alias{getDesign,hadaSeq-method}
\alias{getDesign,mpPar-method}
\alias{getDesign,mpSeq-method}
\alias{getDesign,pbrSeq-method}
\alias{getDesign,rRtbdSeq-method}
\alias{getDesign,rRpbrSeq-method}
\alias{getDesign,rpbrPar-method}
\alias{getDesign,tbdPar-method}
\alias{getDesign,rtbdPar-method}
\alias{getDesign,tbdSeq-method}
\alias{getDesign,udPar-method}
\alias{getDesign,udSeq-method}
\title{Design of a randomization procedure}
\usage{
getDesign(obj)

\S4method{getDesign}{pbrPar}(obj)

\S4method{getDesign}{rarPar}(obj)

\S4method{getDesign}{rarSeq}(obj)

\S4method{getDesign}{abcdPar}(obj)

\S4method{getDesign}{abcdSeq}(obj)

\S4method{getDesign}{bbcdPar}(obj)

\S4method{getDesign}{bbcdSeq}(obj)

\S4method{getDesign}{ebcPar}(obj)

\S4method{getDesign}{bsdPar}(obj)

\S4method{getDesign}{bsdSeq}(obj)

\S4method{getDesign}{chenPar}(obj)

\S4method{getDesign}{chenSeq}(obj)

\S4method{getDesign}{crPar}(obj)

\S4method{getDesign}{crSeq}(obj)

\S4method{getDesign}{ebcSeq}(obj)

\S4method{getDesign}{gbcdPar}(obj)

\S4method{getDesign}{gbcdSeq}(obj)

\S4method{getDesign}{hadaPar}(obj)

\S4method{getDesign}{hadaSeq}(obj)

\S4method{getDesign}{mpPar}(obj)

\S4method{getDesign}{mpSeq}(obj)

\S4method{getDesign}{pbrSeq}(obj)

\S4method{getDesign}{rRtbdSeq}(obj)

\S4method{getDesign}{rRpbrSeq}(obj)

\S4method{getDesign}{rpbrPar}(obj)

\S4method{getDesign}{tbdPar}(obj)

\S4method{getDesign}{rtbdPar}(obj)

\S4method{getDesign}{tbdSeq}(obj)

\S4method{getDesign}{udPar}(obj)

\S4method{getDesign}{udSeq}(obj)
}
\arguments{
\item{obj}{object of the class \code{randSeq} or \code{randPar}.}
}
\value{
the name of the randomization procedure used
}
\description{
Generates a \code{character} vector which specifies the used randomization method
}
