% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randSeq.R
\name{getRandomizationList}
\alias{getRandList}
\alias{getRandomizationList}
\title{Accessor function for the randomization list}
\usage{
getRandList(obj)
}
\arguments{
\item{obj}{object specifying the randomization procedure, i.e. an object of a
class.}
}
\description{
Get the randomization list coded in its groups.
}
\examples{
myPar <- bsdPar(10, 2)
M <- genSeq(myPar, 2)
getRandList(M)
}

