% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublyF.R
\name{hasAllGroups}
\alias{hasAllGroups}
\title{Check function for occurance of all treatment groups in the sequence}
\usage{
hasAllGroups(seq, K)
}
\arguments{
\item{seq}{randomization sequence as inverted matrix}

\item{K}{number of treatment arms}
}
\value{
TRUE if all groups represented, FAlSE otherwise
}
\description{
checks wheather each group has its value comming up at least once in the sequence
}
