% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{K}
\alias{K}
\title{Function returning the total sample size slot of an S4 object}
\usage{
K(obj)
}
\arguments{
\item{obj}{object of class randPar}
}
\description{
Function returning the total sample size slot of an S4 object
}

