% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{method}
\alias{method}
\title{Function returning the method of an S4 object}
\usage{
method(obj)
}
\arguments{
\item{obj}{object inheriting from randPar}
}
\value{
The method of an \code{S4} object
}
\description{
Function returning the method of an S4 object
}
