% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPar.R
\name{randBlocks}
\alias{randBlocks}
\title{Function returning the block slot of an S4 object}
\usage{
randBlocks(obj)
}
\arguments{
\item{obj}{object of class pbrPAr}
}
\description{
Function returning the block slot of an S4 object
}
