% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRand.R
\name{saveRand}
\alias{saveRand}
\title{Saving a randomization lists}
\usage{
saveRand(obj, file = "randList.csv")
}
\arguments{
\item{obj}{object of class \code{randSeq} containing a single randomization sequence.}

\item{file}{A connection, or a character string naming the file to write to.}
}
\value{
Creates a \code{.csv} data in the home folder and saves the randomization list as a column vector.
}
\description{
Saves the parameters of a \code{randSeq} object in a \code{.csv} data
sheet.
}
\seealso{
Other saving functions: \code{\link{saveAssess}}
}
