% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSeq.R
\name{seed}
\alias{seed}
\title{Function returning the allocation seed slot of an object}
\usage{
seed(obj)
}
\arguments{
\item{obj}{object specifying the randomization procedure, see \code{\link{randPar}}
or \code{\link{createParam}}.}
}
\description{
Returns the seed that was either generated at random or user specified.
The seed can be specified for any random operation e.g. genSeq.
}
